Moved Shared to SharedClassLibrary and managed current the repositories and domain (#10)
Moved Shared to SharedClassLibrary, moved IRepo to SharedClassLibrary, moved all repos to the server and made it easier for anyone implementing the proxy repos. I also had to make sure that the right proxy repos are created for the services. I also moved all the domain things in the SharedClassLibrary. I also had to add the Microsoft.Data.SqlClient package to the SharedClassLibrary as some things from Shared needed it